sync fork

+++ zeppelin-web-angular/src/app/pages/workspace/notebook/notebook.component.ts
@@ -19,6 +19,7 @@ import {
   QueryList,
   ViewChildren
 } from '@angular/core';
+import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
 import { isNil } from 'lodash';
 import { Subject } from 'rxjs';
@@ -42,10 +43,11 @@ import {
   NoteStatusService,
   NoteVarShareService,
   SecurityService,
+  ThemeService,
   TicketService
 } from '@zeppelin/services';
 
-import { scrollIntoViewIfNeeded } from '@zeppelin/utility/element';
+import { scrollIntoViewIfNeeded } from '@zeppelin/utility';
 import { NotebookParagraphComponent } from './paragraph/paragraph.component';
 
 @Component({
@@ -55,21 +57,21 @@ import { NotebookParagraphComponent } from './paragraph/paragraph.component';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class NotebookComponent extends MessageListenersManager implements OnInit, OnDestroy {
-  @ViewChildren(NotebookParagraphComponent) listOfNotebookParagraphComponent: QueryList<NotebookParagraphComponent>;
+  @ViewChildren(NotebookParagraphComponent) listOfNotebookParagraphComponent!: QueryList<NotebookParagraphComponent>;
   private destroy$ = new Subject();
-  note: Note['note'];
-  permissions: Permissions;
+  note?: Exclude<Note['note'], undefined>;
+  permissions?: Permissions;
   selectId: string | null = null;
   scrolledId: string | null = null;
   isOwner = true;
   noteRevisions: RevisionListItem[] = [];
-  currentRevision: string;
+  currentRevision?: string;
   collaborativeMode = false;
   revisionView = false;
-  collaborativeModeUsers = [];
-  isNoteDirty = false;
+  collaborativeModeUsers: string[] = [];
+  isNoteDirty: boolean | null = false;
   isShowNoteForms = false;
-  saveTimer = null;
+  saveTimer: ReturnType<typeof setTimeout> | null = null;
   interpreterBindings: InterpreterBindingItem[] = [];
   activatedExtension: 'interpreter' | 'permissions' | 'revisions' | 'hide' = 'hide';
   sidebarWidth = 370;
@@ -86,23 +88,25 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
       this.note = note;
       const { paragraphId } = this.activatedRoute.snapshot.params;
       if (paragraphId) {
-        this.note = this.cleanParagraphExcept(paragraphId);
-        this.initializeLookAndFeel();
+        this.note = this.cleanParagraphExcept(this.note, paragraphId);
+        this.initializeLookAndFeel(this.note);
       } else {
-        this.initializeLookAndFeel();
-        this.getInterpreterBindings();
-        this.getPermissions();
+        this.initializeLookAndFeel(this.note);
+        this.getInterpreterBindings(this.note);
+        this.getPermissions(this.note);
         this.note.config.personalizedMode =
           this.note.config.personalizedMode === undefined ? 'false' : this.note.config.personalizedMode;
       }
-      if (this.note.noteForms && this.note.noteParams) {
+      if (this.note!.noteForms && this.note!.noteParams) {
         this.saveNoteForms({
           formsData: {
-            forms: this.note.noteForms,
-            params: this.note.noteParams
+            forms: this.note!.noteForms,
+            params: this.note!.noteParams
           }
         });
       }
+      this.titleService.setTitle(`${this.note?.name} - Zeppelin`);
+      this.themeService.updateMonacoTheme();
       this.cdr.markForCheck();
     }
   }
@@ -122,7 +126,18 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     if (paragraphId || this.revisionView) {
       return;
     }
-    this.note.paragraphs = this.note.paragraphs.filter(p => p.id !== data.id);
+    if (!this.note) {
+      return;
+    }
+    const definedNote = this.note;
+    const paragraphIndex = definedNote.paragraphs.findIndex(p => p.id === data.id);
+    definedNote.paragraphs = definedNote.paragraphs.filter((p, index) => index !== paragraphIndex);
+    const adjustedCursorIndex =
+      paragraphIndex === definedNote.paragraphs.length ? paragraphIndex - 1 : paragraphIndex + 1;
+    const targetParagraph = this.listOfNotebookParagraphComponent.find((_, index) => index === adjustedCursorIndex);
+    if (targetParagraph) {
+      targetParagraph.focusEditor();
+    }
     this.cdr.markForCheck();
   }
 
@@ -132,21 +147,25 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     if (paragraphId || this.revisionView) {
       return;
     }
-    this.note.paragraphs.splice(data.index, 0, data.paragraph).map(p => {
-      return {
-        ...p,
-        focus: p.id === data.paragraph.id
-      };
-    });
-    this.note.paragraphs = [...this.note.paragraphs];
+    if (!this.note) {
+      return;
+    }
+    const definedNote = this.note;
+    definedNote.paragraphs.splice(data.index, 0, data.paragraph);
+    const paragraphIndex = definedNote.paragraphs.findIndex(p => p.id === data.paragraph.id);
+
+    definedNote.paragraphs[paragraphIndex].focus = true;
     this.cdr.markForCheck();
-    // TODO(hsuanxyz) focus on paragraph
   }
 
   @MessageListener(OP.SAVE_NOTE_FORMS)
   saveNoteForms(data: MessageReceiveDataTypeMap[OP.SAVE_NOTE_FORMS]) {
-    this.note.noteForms = data.formsData.forms;
-    this.note.noteParams = data.formsData.params;
+    if (!this.note) {
+      return;
+    }
+    const definedNote = this.note;
+    definedNote.noteForms = data.formsData.forms;
+    definedNote.noteParams = data.formsData.params;
     this.setNoteFormsStatus();
   }
 
@@ -156,20 +175,23 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     if (isNil(note)) {
       this.router.navigate(['/']).then();
     } else {
-      this.note = data.note;
-      this.initializeLookAndFeel();
+      this.note = note;
+      this.initializeLookAndFeel(this.note);
       this.cdr.markForCheck();
     }
   }
 
   @MessageListener(OP.SET_NOTE_REVISION)
-  setNoteRevision() {
+  setNoteRevision(_data: MessageReceiveDataTypeMap[OP.SET_NOTE_REVISION]) {
     const { noteId } = this.activatedRoute.snapshot.params;
     this.router.navigate(['/notebook', noteId]).then();
   }
 
   @MessageListener(OP.PARAGRAPH_MOVED)
   moveParagraph(data: MessageReceiveDataTypeMap[OP.PARAGRAPH_MOVED]) {
+    if (!this.note) {
+      return;
+    }
     if (!this.revisionView) {
       const movedPara = this.note.paragraphs.find(p => p.id === data.id);
       if (movedPara) {
@@ -181,6 +203,7 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
           // Call when next tick
           setTimeout(() => {
             scrollIntoViewIfNeeded(paragraphComponent.getElement());
+            paragraphComponent.focusEditor();
           });
         }
       }
@@ -195,19 +218,22 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   @MessageListener(OP.PATCH_PARAGRAPH)
-  patchParagraph() {
+  patchParagraph(_data: MessageReceiveDataTypeMap[OP.PATCH_PARAGRAPH]) {
     this.collaborativeMode = true;
     this.cdr.markForCheck();
   }
 
   @MessageListener(OP.NOTE_UPDATED)
   noteUpdated(data: MessageReceiveDataTypeMap[OP.NOTE_UPDATED]) {
+    if (!this.note) {
+      return;
+    }
     if (data.name !== this.note.name) {
       this.note.name = data.name;
     }
     this.note.config = data.config;
     this.note.info = data.info;
-    this.initializeLookAndFeel();
+    this.initializeLookAndFeel(this.note);
     this.cdr.markForCheck();
   }
 
@@ -220,7 +246,11 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
       }
       const { revisionId } = this.activatedRoute.snapshot.params;
       if (revisionId) {
-        this.currentRevision = this.noteRevisions.find(r => r.id === revisionId).message;
+        const revisionItemFound = this.noteRevisions.find(r => r.id === revisionId);
+        if (!revisionItemFound) {
+          throw new Error(`Revision ${revisionId} not found`);
+        }
+        this.currentRevision = revisionItemFound.message;
       } else {
         this.currentRevision = 'Head';
       }
@@ -228,11 +258,16 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     this.cdr.markForCheck();
   }
 
+  onParagraphSearch(term: string) {
+    this.listOfNotebookParagraphComponent.forEach(comp => comp.highlightMatches(term || ''));
+  }
+
   saveParagraph(id: string) {
-    this.listOfNotebookParagraphComponent
-      .toArray()
-      .find(p => p.paragraph.id === id)
-      .saveParagraph();
+    const paragraphFound = this.listOfNotebookParagraphComponent.toArray().find(p => p.paragraph.id === id);
+    if (!paragraphFound) {
+      throw new Error(`Paragraph ${id} not found`);
+    }
+    paragraphFound.saveParagraph();
   }
 
   killSaveTimer() {
@@ -250,15 +285,18 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     }, 10000);
   }
 
-  onParagraphSelect(id: string) {
+  onParagraphSelect(id: string | null) {
     this.selectId = id;
   }
 
-  onParagraphScrolled(id: string) {
+  onParagraphScrolled(id: string | null) {
     this.scrolledId = id;
   }
 
   onSelectAtIndex(index: number) {
+    if (!this.note) {
+      throw new Error(`"note" is not defined. Please check if note data is loaded before calling this method.`);
+    }
     const scopeIndex = Math.min(this.note.paragraphs.length, Math.max(0, index));
     if (this.note.paragraphs[scopeIndex]) {
       this.selectId = this.note.paragraphs[scopeIndex].id;
@@ -275,12 +313,12 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     }
   }
 
-  getInterpreterBindings() {
-    this.messageService.getInterpreterBindings(this.note.id);
+  getInterpreterBindings(note: Exclude<Note['note'], undefined>) {
+    this.messageService.getInterpreterBindings(note.id);
   }
 
-  getPermissions() {
-    this.securityService.getPermissions(this.note.id).subscribe(data => {
+  getPermissions(note: Exclude<Note['note'], undefined>) {
+    this.securityService.getPermissions(note.id).subscribe(data => {
       this.permissions = data;
       this.isOwner = !(
         this.permissions.owners.length && this.permissions.owners.indexOf(this.ticketService.ticket.principal) < 0
@@ -290,23 +328,29 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   get viewOnly(): boolean {
+    if (!this.note) {
+      return false;
+    }
     return this.noteStatusService.viewOnly(this.note);
   }
 
-  initializeLookAndFeel() {
-    this.note.config.looknfeel = this.note.config.looknfeel || 'default';
-    if (this.note.paragraphs && this.note.paragraphs[0]) {
-      this.note.paragraphs[0].focus = true;
+  initializeLookAndFeel(note: Exclude<Note['note'], undefined>) {
+    note.config.looknfeel = note.config.looknfeel || 'default';
+    if (note.paragraphs && note.paragraphs[0]) {
+      note.paragraphs[0].focus = true;
     }
   }
 
-  cleanParagraphExcept(paragraphId) {
-    const targetParagraph = this.note.paragraphs.find(p => p.id === paragraphId);
+  cleanParagraphExcept(note: Exclude<Note['note'], undefined>, paragraphId: string) {
+    const targetParagraph = note.paragraphs.find(p => p.id === paragraphId);
+    if (!targetParagraph) {
+      throw new Error(`Paragraph ${paragraphId} not found`);
+    }
     const config = targetParagraph.config || {};
     config.editorHide = true;
     config.tableHide = false;
     const paragraphs = [{ ...targetParagraph, config }];
-    return { ...this.note, paragraphs };
+    return { ...note, paragraphs };
   }
 
   setAllParagraphTableHide(tableHide: boolean) {
@@ -318,6 +362,9 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   onNoteFormChange(noteParams: DynamicFormParams) {
+    if (!this.note) {
+      throw new Error(`"note" is not defined. Please check if note data is loaded before calling this method.`);
+    }
     this.messageService.saveNoteForms({
       noteParams,
       id: this.note.id
@@ -325,10 +372,16 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   onFormNameRemove(formName: string) {
+    if (!this.note) {
+      throw new Error(`"note" is not defined. Please check if note data is loaded before calling this method.`);
+    }
     this.messageService.removeNoteForms(this.note, formName);
   }
 
   onNoteTitleChange(noteFormTitle: string) {
+    if (!this.note) {
+      throw new Error(`"note" is not defined. Please check if note data is loaded before calling this method.`);
+    }
     this.messageService.updateNote(this.note.id, this.note.name, {
       ...this.note.config,
       noteFormTitle
@@ -336,7 +389,7 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   setNoteFormsStatus() {
-    this.isShowNoteForms = this.note && this.note.noteForms && Object.keys(this.note.noteForms).length !== 0;
+    this.isShowNoteForms = !!this.note && this.note.noteForms && Object.keys(this.note.noteForms).length !== 0;
     this.cdr.markForCheck();
   }
 
@@ -352,15 +405,17 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
   }
 
   constructor(
-    private activatedRoute: ActivatedRoute,
     public messageService: MessageService,
+    protected ngZService: NgZService,
+    private activatedRoute: ActivatedRoute,
     private cdr: ChangeDetectorRef,
     private noteStatusService: NoteStatusService,
     private noteVarShareService: NoteVarShareService,
     private ticketService: TicketService,
     private securityService: SecurityService,
     private router: Router,
-    protected ngZService: NgZService
+    private titleService: Title,
+    private themeService: ThemeService
   ) {
     super(messageService);
   }
@@ -376,14 +431,9 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
         this.onParagraphSelect(id);
         this.onParagraphScrolled(id);
       });
-    this.activatedRoute.params
-      .pipe(
-        takeUntil(this.destroy$),
-        distinctUntilKeyChanged('noteId')
-      )
-      .subscribe(() => {
-        this.noteVarShareService.clear();
-      });
+    this.activatedRoute.params.pipe(takeUntil(this.destroy$), distinctUntilKeyChanged('noteId')).subscribe(() => {
+      this.noteVarShareService.clear();
+    });
     this.activatedRoute.params.pipe(takeUntil(this.destroy$)).subscribe(param => {
       const { noteId, revisionId } = param;
       if (revisionId) {
@@ -413,5 +463,6 @@ export class NotebookComponent extends MessageListenersManager implements OnInit
     this.saveNote();
     this.destroy$.next();
     this.destroy$.complete();
+    this.titleService.setTitle('Zeppelin');
   }
 }
